Feat: 쿼리DSL 기능 추가

+++ gathering/src/main/java/com/sparta/moim/gathering/presentation/dto/response/SearchGatheringResponse.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public record SearchGatheringResponse(
     List<SearchGatheringListResponse> gatherings,
-    int total,
+    long total,
     int page,
     int content
 ) {
